tests/slicing: fix C library related warnings
Namely, -Wimplicit-function-declaration and
-Wincompatible-library-redeclaration.

+++ tests/slicing/sources/malloc-redef.c
@@ -1,6 +1,8 @@
+#include <stddef.h>
+
 int x;
 
-void *malloc(int i) {
+void *malloc(size_t i) {
     if (i == sizeof(int))
         return &x;
     else
